refactor: comprehensive audit - fix 30+ issues across core, network, and UI
CRITICAL fixes (crash/security):
- Fix NULL deref in MinecraftAccount::refresh() when non-Local account
- Fix missing return after ContentGoneError in Yggdrasil auth
- Fix Task.h progress overflow (int → qint64)
- Fix FileSink::abort() crash on null m_output_file
- Fix SSL errors never rejected (MITM prevention in Download + Yggdrasil)
- Fix dynamic_cast without null checks in MainWindow::updateToolsMenu
- Fix CustomMessageBox memory leak (WA_DeleteOnClose)
- Fix connection leak in updateToolsMenu

HIGH fixes (wrong behavior):
- Fix decideAccount() duplicate dead call → fallback to accounts->at(0)
- Fix proxy settings never applied at startup
- Fix instanceSet not cleaned on instance removal
- Fix m_setupWizard dangling pointer after wizard finishes
- Fix redirect loop protection (MAX_REDIRECTS=10)
- Fix URL scheme validation for redirects (http/https only)
- Fix HttpMetaCache::Load() aborting on single invalid entry
- Fix missing return after emitFailed in AssetUpdateTask
- Fix WorldListPage dead MCEdit code and typo

MEDIUM fixes (code quality):
- Fix unbounded ByteArraySink memory (100MB limit)
- Fix SkinUpload toLocal8Bit → toUtf8 encoding
- Add SequentialTask canAbort/abort support
- Add ComponentUpdateTask recursion depth guard
- Fix FMLLibrariesTask null check for downloadJob
- Fix LauncherPage raw QTextCharFormat → value member
- Fix InstanceWindow stylesheet hack → proper polish
- Remove unnecessary Q_OBJECT from OrSetting
- Fix AuthRequest deleteLater ordering
- Fix AboutDialog MultiMC → BuildConfig.LAUNCHER_NAME
- Fix LauncherPage displayName() translatable
- Fix PageContainer deprecated setMargin → setContentsMargins

+++ launcher/tasks/Task.h
@@ -104,7 +104,7 @@ private:
     QStringList m_Warnings;
     QString m_failReason = "";
     QString m_status;
-    int m_progress = 0;
-    int m_progressTotal = 100;
+    qint64 m_progress = 0;
+    qint64 m_progressTotal = 100;
 };
